feat: implement useMCPSelect hook for managing MCP selection state, refactor MCPSelect component to utilize new hook
