+++ client/src/hooks/Plugins/index.ts
@@ -1,6 +1,7 @@
+export * from './useMCPSelect';
+export * from './useToolToggle';
 export { default as useAuthCodeTool } from './useAuthCodeTool';
 export { default as usePluginInstall } from './usePluginInstall';
 export { default as useCodeApiKeyForm } from './useCodeApiKeyForm';
 export { default as useSearchApiKeyForm } from './useSearchApiKeyForm';
 export { default as usePluginDialogHelpers } from './usePluginDialogHelpers';
-export { useToolToggle } from './useToolToggle';
